Refactored vae based detection

See merge request benson-mbani/deepsea-fauna-detection!2

+++ src/custom_object_detection/tf_object_detection_utilities.py
@@ -11,7 +11,9 @@ import io
 from object_detection.utils import label_map_util
 from object_detection.utils import visualization_utils as viz_utils
 import numpy as np
+from zipfile import ZipFile
 import matplotlib.pyplot as plt
+from concurrent.futures import ThreadPoolExecutor
 
 
 def pre_process_csv_with_labels_to_have_one_row_per_image(path_to_csv_with_labels):
@@ -123,6 +125,28 @@ def create_tf_example_from_a_row_of_csv_with_labels(row_of_csv_as_named_tuple):
     return tf_example
 
 
+# def generate_tfrecords_from_csv_with_labels(path_to_csv_with_labels):
+#     '''
+#     Given a csv with labels from image viewer, process its columns to generate trecord files
+    
+#     '''
+#     processed_df = pre_process_csv_with_labels_to_have_one_row_per_image(path_to_csv_with_labels)
+    
+#     rows_of_csv_as_named_tuple = processed_df.itertuples()
+    
+#     tf_examples = [create_tf_example_from_a_row_of_csv_with_labels(row_of_csv) for row_of_csv in rows_of_csv_as_named_tuple]
+
+#     random.shuffle(tf_examples)
+    
+#     number_of_train_examples = int(0.9 * len(tf_examples))
+    
+#     tf_train_examples = tf_examples[:number_of_train_examples]
+    
+#     tf_val_examples = tf_examples[number_of_train_examples:]
+    
+#     return tf_train_examples, tf_val_examples
+
+
 def generate_tfrecords_from_csv_with_labels(path_to_csv_with_labels):
     '''
     Given a csv with labels from image viewer, process its columns to generate trecord files
@@ -133,16 +157,8 @@ def generate_tfrecords_from_csv_with_labels(path_to_csv_with_labels):
     rows_of_csv_as_named_tuple = processed_df.itertuples()
     
     tf_examples = [create_tf_example_from_a_row_of_csv_with_labels(row_of_csv) for row_of_csv in rows_of_csv_as_named_tuple]
-
-    random.shuffle(tf_examples)
     
-    number_of_train_examples = int(0.9 * len(tf_examples))
-    
-    tf_train_examples = tf_examples[:number_of_train_examples]
-    
-    tf_val_examples = tf_examples[number_of_train_examples:]
-    
-    return tf_train_examples, tf_val_examples
+    return tf_examples
 
 
 def download_config_file(directory_to_save_config_file, config_file_source_url):
@@ -183,7 +199,9 @@ def download_checkpoint_for_pretraining(detection_checkpoint_url, directory_to_s
     
     pipeline_config_file = directory_to_save_checkpoint / f'{unpacked_directory_name}/pipeline.config'
     
-    shutil.copy2(pipeline_config_file, directory_to_save_checkpoint/'my_model_template.config')
+    if pipeline_config_file.exists():
+    
+        shutil.copy2(pipeline_config_file, directory_to_save_checkpoint/'my_model_template.config')
     
     Path(file_save_path).unlink()
     
@@ -191,20 +209,24 @@ def download_checkpoint_for_pretraining(detection_checkpoint_url, directory_to_s
 
 
 
-def create_train_val_input_tfrecords(path_to_csv_with_labels, path_to_label_map, path_to_train_tfrecord_file, path_to_validation_tfrecord_file):
+def create_train_val_input_tfrecords(path_to_train_csv_with_labels, path_to_val_csv_with_labels, path_to_label_map, path_to_train_tfrecord_file, path_to_validation_tfrecord_file):
        
     train_writer = tf.io.TFRecordWriter(str(path_to_train_tfrecord_file))
     
     validation_writer = tf.io.TFRecordWriter(str(path_to_validation_tfrecord_file))
     
-    tf_train_examples, tf_val_examples = generate_tfrecords_from_csv_with_labels(path_to_csv_with_labels)
+    tf_train_examples = generate_tfrecords_from_csv_with_labels(path_to_train_csv_with_labels)
     
+    tf_val_examples = generate_tfrecords_from_csv_with_labels(path_to_val_csv_with_labels)
+    
+    print(f'Writing {len(tf_train_examples)} training examples ...')
     for tf_example in tf_train_examples:
         
         train_writer.write(tf_example.SerializeToString())
 
     train_writer.close()
     
+    print(f'Writing {len(tf_val_examples)} validation examples ...')
     for tf_example in tf_val_examples:
         
         validation_writer.write(tf_example.SerializeToString())
@@ -285,6 +307,103 @@ def visualize_detections(image_tensor, detections, category_index, score_thresho
     ax.set_yticks([])
     plt.savefig(Path(directory_to_save_detection_figures) / f'{figname}.png', dpi=300, bbox_inches='tight')
     plt.close()
+    
+    
+    
+def generate_matrix_of_detections(detections, score_threshold, file_path):
+    '''
+    Given a detection instance, process it to generate matrix of detection results
+    
+    '''
+    
+    scores  = detections['detection_scores'][0].numpy().reshape(-1,1)
+    classes = detections['detection_classes'][0].numpy().reshape(-1,1)
+    boxes = detections['detection_boxes'][0].numpy() * [1120, 1680, 1120, 1680]
+    file_paths = np.array([str(Path(file_path))]*len(scores)).reshape(-1,1)
+    file_names = np.array([str(Path(file_path).stem)]*len(scores)).reshape(-1,1)
+
+    detection_matrix = np.hstack([file_names, scores, classes, boxes, file_paths])
+
+    selector = (scores > score_threshold).squeeze()
+
+    selected_detection_matrix = np.compress(selector,detection_matrix, axis=0)
+    
+    return selected_detection_matrix
+
+
+
+def append_geospatial_coordinates_to_master_detection_summary_table(master_dataframe_summarizing_detections):
+    '''
+    Append geospatial coordinates to master dataframe summarizing detections before saving to disk
+    
+    '''
+    path_to_table_with_georeferenced_coords_for_all_photos = Path.cwd().parents[0]/'reports/auxilliary_data/georeferenced_photo_coordinates.zip'
+    
+    with ZipFile(path_to_table_with_georeferenced_coords_for_all_photos) as myzip:
+        
+        with myzip.open('georeferenced_photo_coordinates.csv') as myfile:
+            
+            table_with_georeferenced_coords_for_all_photos = pd.read_csv(myfile)
+            
+    table_with_georeferenced_coords_for_all_photos['parent_image_name'] = table_with_georeferenced_coords_for_all_photos.Name.map(lambda x: x.split('.')[0])
+            
+    georeferenced_master_dataframe_summarizing_detections = pd.merge(master_dataframe_summarizing_detections, table_with_georeferenced_coords_for_all_photos, how='left', on='parent_image_name')
+    
+    return georeferenced_master_dataframe_summarizing_detections
+
+
+def save_georeferenced_detection_results_as_csv(complete_detection_matrix, label_map_path, detection_results_csv_file_name):
+    '''
+    Georeference detection results and save to disk
+    
+    '''
+    columns = ['parent_image_name', 'score', 'object_class_id', 'ymin', 'xmin', 'ymax', 'xmax', 'parent_image_path']
+
+    detection_matrix_dataframe = pd.DataFrame(complete_detection_matrix, columns=columns)
+    
+    georeferenced_detection_matrix_dataframe = append_geospatial_coordinates_to_master_detection_summary_table(detection_matrix_dataframe)
+    
+    label_map = label_map_util.load_labelmap(label_map_path)
+
+    label_map_dict = label_map_util.get_label_map_dict(label_map, use_display_name=False)
+    
+    label_map_dict_reversed = {v:k for k,v in label_map_dict.items()}
+    
+    object_class_name = georeferenced_detection_matrix_dataframe.object_class_id.map(lambda x:label_map_dict_reversed[int(float(x))])
+    
+    georeferenced_detection_matrix_dataframe['object_class_name'] = object_class_name
+    
+    georeferenced_detection_matrix_dataframe = georeferenced_detection_matrix_dataframe.loc[:,['parent_image_name', 'object_class_name', 'object_class_id', 'score', 'ymin', 'xmin', 'ymax', 'xmax', 'dive', 'time_stamp', 'Latitude', 'Longitude', 'recoded_classification', 'license_area', 'activity', 'parent_image_path']].rename(columns={'recoded_classification':'seafloor_classification'})
+
+    georeferenced_detection_matrix_dataframe.to_csv(detection_results_csv_file_name, index=False)
+    
+    
+    
+def sample_ground_truth_images_for_annotation(directory_with_all_parent_images, directory_to_save_samples, path_to_annotations_table, n_samples):
+    '''
+    Sample images with the most detections for manual annotation and error reporting
+    
+    '''
+    directory_to_save_samples = Path(directory_to_save_samples)
+    
+    path_to_annotations_table = Path(path_to_annotations_table)
+    
+    directory_with_all_parent_images = Path(directory_with_all_parent_images)
+    
+    df = pd.read_csv(path_to_annotations_table)
+    
+    number_of_detections_df = df.groupby('image_name').size().to_frame(name='number_of_detections').reset_index()
+    
+    sorted_number_of_detections_df = number_of_detections_df.sort_values(by='number_of_detections', ascending=False, ignore_index=True)
+    
+    file_names_to_sample = sorted_number_of_detections_df.loc[:n_samples, 'image_name']
+    
+    file_paths_to_sample = [(directory_with_all_parent_images / name) for name in file_names_to_sample]
+    
+    with ThreadPoolExecutor() as executor:
+        
+        [executor.submit(shutil.copy2, fp, directory_to_save_samples) for fp in file_paths_to_sample]
+
